fix: Interpolation factor (#20)
Both `RollbackSynchronizer` and `TickInterpolator` used `df` to
interpolate - the idea was to calculate a factor based on how much is
left from the tick and interpolate between the *current* state and the
target state, instead of the *previous* tick state and the target. This
promised some leeway in masking simulation artifacts, but `df` turned
out to be numerically unstable.

The fix is to simply interpolate between the previous and target tick
state without the `df` wizardry.