fix: Jerky interpolation
RollbackSynchronizer always used a state from history as a starting
point for interpolation. If that state changed as we go ( i.e. we got
an override from the server ), the interpolation jumped, causing a jerk
in the motion.
The fix is to always use the state captured before the tick loop.