fix: Powerup effects stay even after effect ends (#44)
Issue only happened in Release builds, worked fine in Debug. For some
reason, in Release, the `free()` method wasn't found in the Effect
class, basically removing all effect classes. This wasn't apparent
during gameplay, as the server simulated the effects and synced to the
clients.
Switching to `queue_free()` fixed the issue.

Closes #43