fix(game): Players getting points for hitting themselves
In some cases, the players can get caught up in their own blast zones.
When they fall after this, the score manager will award an extra point
to whoever hit them last - which is the player itself.
This is fixed by adding some extra checks, and an additional error
message in case this happens.

Refs #48