fix: TickInterpolator glitch on state switch (#301)
At the end of the tick loop, `TickInterpolator` catches the new state
and starts interpolating towards it. However, between
`NetworkTime.after_tick_loop` and the next `TickInterpolator._process`
callback, there may be a single frame where the interpolated node(s) are
still in the target state, and then get set back to the starting state
once the interpolation kicks in. This leads to a periodic glitching.

To fix this, `TickInterpolator` always applies the starting state once
it's captured, in `NetworkTime.after_tick_loop`.

Closes #287