feat: Nudge Time Process (#314)
Implements a time-sync algorithm inspired by [NTPv4] to get a reliable
idea of the current time at the host. Based on the host's clock, an
internal reference clock is kept and adjusted to stay in sync. A
separate, simulation clock is used to drive the network tick loop. The
simulation clock is constantly sped up / slowed down slightly, to stay
in sync with the reference clock, and by extension, with the host's
clock.

This also deprecates some settings and properties, including the three
separate tick concepts ( remote tick, local tick, tick ) along with
their time counterparts. With this new system, only `tick` and `time`
exists, since they're kept in sync with the host at all times.

[NTPv4]: https://datatracker.ietf.org/doc/html/rfc5905

Closes #176