refactor: NetfoxLogger no longer depends on netfox (#336)
Directly referencing `NetworkTime` and `NetworkEvents` in
`_NetfoxLogger` means that `netfox.internals` depends on `netfox`. This
creates both a circular dependency, and an unnecessary dependency for
`netfox.noray`. The noray integration doesn't need netfox itself to work
properly, and should be kept usable *without* netfox itself.

Instead, tags can be registered in `_NetfoxLogger`, which are functions
that provide information for all log messages. Tags are ordered, to
ensure the same log message format, regardless of the order the
registering may run their init functions.