fix: Avoid NaNs at very low latencies (#469)
`NetworkTimeSynchronizer` samples the host's clock and takes a weighted
average of the clock offsets, based on the roundtrip time. If the
roundtrip time is very close to zero, the weight itself will zero, thus
resulting in a division by zero. This value is then used to adjust the
clock, setting it to NaN. This blocks the network tick loop.

To fix this, a check is added that will avoid the weighted average if
the RTT is too low, and calculates a regular average instead.