feat: Track and expose latest confirmed tick (#518)
Allows NetworkRollback to track and expose the latest tick for which
each node with a RollbackSynchronizer has received input submissions.

Useful for scenarios where you need to ensure all input ticks are
submitted before proceeding with actions that are difficult to reverse,
such as character death, visual effects, or audio triggers.

Closes #470

---------

Co-authored-by: Tamás Gálffy <ezittgtx@gmail.com>