fix: RewindableAction to handle later ticks (#534)
When modifying a rewindable action from outside of its parent, it may
set a tick in one loop, but not in the next one. For example, the player
hits a spike, and the spike sets the RewindableAction for tick 15. On
the next loop, the player doesn't hit the spike on tick 15, but is also
not simulated for tick 15 since it doesn't have input.

This results in TicksetSerializer's assert failing, because there's an
active tick past the last set one.

Instead, RewindableAction now serializes the last set tick, even if it
was not set in the current loop. TicksetSerializer will not serialize
ticks outside of the specified range.

Fixes #532